stethoscope & tomograph: backing-out changeset 8acc8d2b1ffb
No more need for down-casts; the code had been fixed in changesets
f4ddd6705bf0 & c1796d6e9327 by using matching types (size & signedness).

+++ clients/mapiclient/stethoscope.c
@@ -291,7 +291,7 @@ main(int argc, char **argv)
 		if (*response) {
 			if (debug)
 				printf("LASTLINE:%s", response);
-			len = (int) strlen(response);
+			len = strlen(response);
 			strncpy(buf, response, len + 1);
 		} else
 			len = 0;
